move physics types into types
